Fix __repr__ for datetime redefinition
repr() of a datetime.datetime object got 'new_datetime' as the class name,
this led to errors with the SimpleCache module that tried to restore the
representation but could not find the new_datetime class.

This fix changes the __repr__ function of the new class to show
'datetime.datetime' as the classes name.